keymap:
Change `w` (Word) to `t` (Token)
Change `W` (Subword) to `w` (Word)
Change `T` (Raise) to `^`
