+++ docs/docs/normal-mode/selection-modes/regex-based.mdx
@@ -1,7 +1,8 @@
 ---
 sidebar_position: 3
 ---
-import {TutorialFallback} from '@site/src/components/TutorialFallback'; 
+
+import {TutorialFallback} from '@site/src/components/TutorialFallback';
 
 # Regex-based
 
@@ -21,25 +22,25 @@ Keybinding: `E`
 
 Same as [Line](#line), however, leading whitespaces are selected, and trailing whitespaces, including newline characters are also selected.
 
-## Subword
+## Word
 
-Keybinding: `W`
+Keybinding: `w`
 
-This selects subwords, even if these words are not separated by spaces.
+This selects word within a token.
 
 For example, `myOatPepperBanana` consists of 4 short words, namely: `my`, `Oat`, `Pepper` and `Banana`.
 
 This is useful for renaming identifiers, especially if we only want to change a single word of the name. [^1]
 
-<TutorialFallback filename="subword"/>
-
-## Word
-
-Keybinding: `w`
-
-Like [Subword](#subword), but it treats each word as a sequence of alphanumeric characters (including `-` and `_`).
-
 <TutorialFallback filename="word"/>
+
+## Token
+
+Keybinding: `t`
+
+Like [Word](#word), but it treats each unit as a sequence of alphanumeric characters (including `-` and `_`).
+
+<TutorialFallback filename="token"/>
 
 [^1]: This is possible because even Prompt is an editor, so the Word mode also works there. See [Core Concepts](../../core-concepts.md#2-every-component-is-a-buffereditor)
