+++ src/recipes.rs
@@ -11,19 +11,207 @@ pub(crate) fn recipe_groups() -> Vec<RecipeGroup> {
         showcase(),
         syntax_node(),
         RecipeGroup {
-            filename: "expand",
-            recipes: [Recipe {
-                description: "Expand to nearest brackets/quotes",
-                content: "hello '{World Foo} bar'".trim(),
-                file_extension: "md",
-                prepare_events: keys!("w n"),
-                events: keys!("t t t t"),
-                expectations: &[CurrentSelectedTexts(&["'{World Foo} bar'"])],
-                terminal_height: None,
-                similar_vim_combos: &[],
-                only: false,
-            }]
-            .to_vec(),
+            filename: "delete",
+            recipes: [
+                Recipe {
+                    description: "Delete lines",
+                    content: "
+To be, or not to be?
+That, is the question.
+Why?
+"
+                    .trim(),
+                    file_extension: "md",
+                    prepare_events: &[],
+                    events: keys!("e d d"),
+                    expectations: &[CurrentSelectedTexts(&["Why?"])],
+                    terminal_height: None,
+                    similar_vim_combos: &["d d", "d j"],
+                    only: false,
+                },
+                Recipe {
+                    description: "Delete word (forward)",
+                    content: "camelCase snake_case".trim(),
+                    file_extension: "md",
+                    prepare_events: &[],
+                    events: keys!("w d d"),
+                    expectations: &[CurrentSelectedTexts(&["snake"])],
+                    terminal_height: None,
+                    similar_vim_combos: &[],
+                    only: false,
+                },
+                Recipe {
+                    description: "Delete token (backward)",
+                    content: "camelCase snake_case PascalCase".trim(),
+                    file_extension: "md",
+                    prepare_events: &[],
+                    events: keys!("t . D D"),
+                    expectations: &[CurrentSelectedTexts(&["camelCase"])],
+                    terminal_height: None,
+                    similar_vim_combos: &[],
+                    only: false,
+                }
+            ].to_vec(),
+        },
+        RecipeGroup {
+            filename: "extend",
+            recipes: [
+                Recipe {
+                    description: "Extend selection (Word)",
+                    content: "foo bar spam pi".trim(),
+                    file_extension: "md",
+                    prepare_events: &[],
+                    events: keys!("w v l l h"),
+                    expectations: &[CurrentSelectedTexts(&["foo bar"])],
+                    terminal_height: None,
+                    similar_vim_combos: &[],
+                    only: false,
+                },
+                Recipe {
+                    description: "Extend selection (Syntax Node)",
+                    content: "[{\"x\": 123}, true, {\"y\": {}}]".trim(),
+                    file_extension: "json",
+                    prepare_events: keys!("z l"),
+                    events: keys!("s v n n b"),
+                    expectations: &[CurrentSelectedTexts(&["{\"x\": 123}, true"])],
+                    terminal_height: None,
+                    similar_vim_combos: &[],
+                    only: false,
+                },
+                Recipe {
+                    description: "Extend selection (Switch Direction)",
+                    content: "foo bar spam baz tim".trim(),
+                    file_extension: "md",
+                    prepare_events: &[],
+                    events: keys!("w l l v l o h o l"),
+                    expectations: &[CurrentSelectedTexts(&["bar spam baz tim"])],
+                    terminal_height: None,
+                    similar_vim_combos: &[],
+                    only: false,
+                }
+            ].to_vec(),
+        },
+        RecipeGroup {
+            filename: "surround",
+            recipes: [
+                Recipe {
+                    description: "Surround",
+                    content: "hello world".trim(),
+                    file_extension: "md",
+                    prepare_events: &[],
+                    events: keys!("e v s ("),
+                    expectations: &[CurrentComponentContent("(hello world)")],
+                    terminal_height: None,
+                    similar_vim_combos: &[],
+                    only: false,
+                },
+                Recipe {
+                    description: "Delete Surround",
+                    content: "(hello world)".trim(),
+                    file_extension: "md",
+                    prepare_events: keys!("z l"),
+                    events: keys!("v d ("),
+                    expectations: &[CurrentComponentContent("hello world")],
+                    terminal_height: None,
+                    similar_vim_combos: &[],
+                    only: false,
+                },
+                Recipe {
+                    description: "Change Surround",
+                    content: "(hello world)".trim(),
+                    file_extension: "md",
+                    prepare_events: keys!("z l"),
+                    events: keys!("v c ( {"),
+                    expectations: &[CurrentComponentContent("{hello world}")],
+                    terminal_height: None,
+                    similar_vim_combos: &[],
+                    only: false,
+                },
+                Recipe {
+                    description: "Select Inside Enclosures",
+                    content: "(hello world)".trim(),
+                    file_extension: "md",
+                    prepare_events: keys!("z l"),
+                    events: keys!("v i ("),
+                    expectations: &[CurrentSelectedTexts(&["hello world"])],
+                    terminal_height: None,
+                    similar_vim_combos: &["v i ("],
+                    only: false,
+                },
+                Recipe {
+                    description: "Select Around Enclosures",
+                    content: "(hello world)".trim(),
+                    file_extension: "md",
+                    prepare_events: keys!("z l"),
+                    events: keys!("v a ("),
+                    expectations: &[CurrentSelectedTexts(&["(hello world)"])],
+                    terminal_height: None,
+                    similar_vim_combos: &["v a ("],
+                    only: false,
+                }
+            ].to_vec(),
+        },
+        RecipeGroup {
+            filename: "replace-cut",
+            recipes: [
+                Recipe {
+                    description: "Replace cut",
+                    content: "
+foo(bar, 1 + 1, spam)
+3 * 10
+"
+                    .trim(),
+                    file_extension: "js",
+                    prepare_events: &[],
+                    events: keys!("s y n R j n j n r"),
+                    expectations: &[CurrentComponentContent( "foo(bar, 1 + 1, spam)\nfoo(bar, 3 * 10, spam)", )],
+                    terminal_height: None,
+                    similar_vim_combos: &["p"],
+                    only: false,
+                }
+            ].to_vec(),
+        },
+        RecipeGroup {
+            filename: "paste",
+            recipes: [
+                Recipe {
+                    description: "Paste with automatic gap insertion (Line)",
+                    content: "
+foo bar
+spam baz
+"
+                    .trim(),
+                    file_extension: "md",
+                    prepare_events: &[],
+                    events: keys!("e y p"),
+                    expectations: &[CurrentComponentContent("foo bar\nfoo bar\nspam baz")],
+                    terminal_height: None,
+                    similar_vim_combos: &[],
+                    only: false,
+                },
+                Recipe {
+                    description: "Paste with automatic gap insertion (Syntax Node)",
+                    content: "function foo(bar: Bar, spam: Spam) {}",
+                    file_extension: "ts",
+                    prepare_events: keys!("/ b a r enter"),
+                    events: keys!("s y p"),
+                    expectations: &[CurrentComponentContent("function foo(bar: Bar, bar: Bar, spam: Spam) {}")],
+                    terminal_height: None,
+                    similar_vim_combos: &[],
+                    only: false,
+                },
+                Recipe {
+                    description: "Paste without automatic gap insertion",
+                    content: "foo bar".trim(),
+                    file_extension: "md",
+                    prepare_events: &[],
+                    events: keys!("e y $ p"),
+                    expectations: &[CurrentComponentContent("foo barfoo bar")],
+                    terminal_height: None,
+                    similar_vim_combos: &[],
+                    only: false,
+                }
+            ].to_vec(),
         },
         RecipeGroup {
             filename: "join",
@@ -70,10 +258,10 @@ def foo():
             filename: "exchange",
             recipes: [
                 Recipe {
-                    description: "Exchange sibling node",
+                    description: "Exchange sibling node (JSON)",
                     content: "[{\"x\": 123}, true, {\"y\": {}}]".trim(),
                     file_extension: "json",
-                    prepare_events: keys!("^ l"),
+                    prepare_events: keys!("z l"),
                     events: keys!("s x n n"),
                     expectations: &[
                         CurrentSelectedTexts(&["{\"x\": 123}"]),
@@ -84,11 +272,11 @@ def foo():
                     only: false,
                 },
                 Recipe {
-                    description: "Exchange sibling node",
+                    description: "Exchange sibling node (XML)",
                     content: "<x><y>foo</y><div/></x>".trim(),
                     file_extension: "xml",
-                    prepare_events: keys!("^ l l l"),
-                    events: keys!("s b x n"),
+                    prepare_events: keys!("z l l l"),
+                    events: keys!("s j x n"),
                     expectations: &[
                         CurrentSelectedTexts(&["<y>foo</y>"]),
                         CurrentComponentContent("<x><div/><y>foo</y></x>"),
@@ -231,10 +419,10 @@ fn foo() {
             .to_vec(),
         },
         RecipeGroup {
-            filename: "subword",
+            filename: "word",
             recipes: [
                 Recipe {
-                    description: "Subword: up/down/left/right movement",
+                    description: "Word: up/down/left/right movement",
                     content: "
 HTTPNetwork 88 kebab-case 
 snake_case 99 PascalCase
@@ -242,33 +430,33 @@ snake_case 99 PascalCase
                     .trim(),
                     file_extension: "md",
                     prepare_events: &[],
-                    events: keys!("W l l l l l j h h h h h k"),
+                    events: keys!("w l l l l l j h h h h h k"),
                     expectations: &[CurrentSelectedTexts(&["HTTP"])],
                     terminal_height: None,
                     similar_vim_combos: &[],
                     only: false,
                 },
                 Recipe {
-                    description: "Subword: next/previous movement (skip symbols)",
+                    description: "Word: next/previous movement (skip symbols)",
                     content: "
 camelCase , kebab-case snake_case
 "
                     .trim(),
                     file_extension: "md",
                     prepare_events: &[],
-                    events: keys!("W n n n n n N N N N N"),
+                    events: keys!("w n n n n n b b b b b"),
                     expectations: &[CurrentSelectedTexts(&["camel"])],
                     terminal_height: None,
                     similar_vim_combos: &[],
                     only: false,
                 },
                 Recipe {
-                    description: "Subword: first/last movement",
+                    description: "Word: first/last movement",
                     content: "hello HTTPNetworkRequestMiddleware world"
                     .trim(),
                     file_extension: "md",
                     prepare_events: &[],
-                    events: keys!("W l . ,"),
+                    events: keys!("w l . ,"),
                     expectations: &[CurrentSelectedTexts(&["HTTP"])],
                     terminal_height: None,
                     similar_vim_combos: &[],
@@ -278,10 +466,10 @@ camelCase , kebab-case snake_case
             .to_vec(),
         },
         RecipeGroup {
-            filename: "word",
+            filename: "Token",
             recipes: [
                 Recipe {
-                    description: "Word: up/down/left/right movement",
+                    description: "Token: up/down/left/right movement",
                     content: "
 camelCase ,  kebab-case 
 snake_case + PascalCase
@@ -289,14 +477,14 @@ snake_case + PascalCase
                     .trim(),
                     file_extension: "md",
                     prepare_events: &[],
-                    events: keys!("w l l j h h k"),
+                    events: keys!("t l l j h h k"),
                     expectations: &[CurrentSelectedTexts(&["camelCase"])],
                     terminal_height: None,
                     similar_vim_combos: &[],
                     only: false,
                 },
                 Recipe {
-                    description: "Word: next/previous movement (skip symbols)",
+                    description: "Token: next/previous movement (skip symbols)",
                     content: "
 camelCase , kebab-case 
 snake_case + PascalCase
@@ -304,18 +492,18 @@ snake_case + PascalCase
                     .trim(),
                     file_extension: "md",
                     prepare_events: &[],
-                    events: keys!("w n n n N N N"),
+                    events: keys!("t n n n b b b"),
                     expectations: &[CurrentSelectedTexts(&["camelCase"])],
                     terminal_height: None,
                     similar_vim_combos: &[],
                     only: false,
                 },
                 Recipe {
-                    description: "Word: first/last movement (sentence boundary)",
+                    description: "Token: first/last movement (sentence boundary)",
                     content: "The quick brown fox jumps! Over the lazy dog today...".trim(),
                     file_extension: "md",
                     prepare_events: &[],
-                    events: keys!("w . n . , N ,"),
+                    events: keys!("t . n . , b ,"),
                     expectations: &[CurrentSelectedTexts(&["The"])],
                     terminal_height: None,
                     similar_vim_combos: &[],
@@ -332,7 +520,7 @@ snake_case + PascalCase
                     content: "count > 0 ? x + 2 : y / z".trim(),
                     file_extension: "js",
                     prepare_events: keys!("/ x enter"),
-                    events: keys!("s T"),
+                    events: keys!("s ^"),
                     expectations: &[
                         CurrentSelectedTexts(&["x + 2"]),
                         CurrentComponentContent("x + 2"),
@@ -353,7 +541,7 @@ snake_case + PascalCase
                         .trim(),
                     file_extension: "js",
                     prepare_events: keys!("/ < c h i l d enter"),
-                    events: keys!("s T"),
+                    events: keys!("s ^"),
                     expectations: &[
                         CurrentSelectedTexts(&["<Child x={y}/>"]),
                         CurrentComponentContent("<GParent>\n    <Child x={y}/>\n</GParent>"),
@@ -371,7 +559,7 @@ app.post('/admin', () => {
                     .trim(),
                     file_extension: "js",
                     prepare_events: keys!("/ r o u t e r enter"),
-                    events: keys!("s T T"),
+                    events: keys!("s ^ ^"),
                     expectations: &[
                         CurrentSelectedTexts(&["router.route(foo, bar)"]),
                         CurrentComponentContent("app.post('/admin', () => router.route(foo, bar))"),
@@ -385,7 +573,7 @@ app.post('/admin', () => {
                     content: r#"{"hello": {"world": [123], "foo": null}}"#.trim(),
                     file_extension: "js",
                     prepare_events: keys!("/ 1 2 3 enter"),
-                    events: keys!("s T T"),
+                    events: keys!("s ^ ^"),
                     expectations: &[
                         CurrentSelectedTexts(&["123"]),
                         CurrentComponentContent(r#"{"hello": 123}"#),
@@ -434,7 +622,7 @@ fn foo() {
                     .trim(),
                     file_extension: "rs",
                     prepare_events: &[],
-                    events: keys!("e j t q e"),
+                    events: keys!("e j v i { q e"),
                     expectations: &[CurrentSelectedTexts(&["bar();", "spam();", "baz();"])],
                     terminal_height: Some(7),
                     similar_vim_combos: &[],
@@ -449,7 +637,7 @@ foo ha"
                         .trim(),
                     file_extension: "rs",
                     prepare_events: &[],
-                    events: keys!("e v j q / f o o enter w q W q r - enter"),
+                    events: keys!("e v j q / f o o enter t q w q r - enter"),
                     expectations: &[CurrentSelectedTexts(&[
                         "foo", "da", "foo", "baz", "foo", "yo",
                     ])],
@@ -483,7 +671,7 @@ foo ha"
                     .trim(),
                     file_extension: "md",
                     prepare_events: &[],
-                    events: keys!("w . q h h"),
+                    events: keys!("t . q h h"),
                     expectations: &[
                         CurrentSelectedTexts(&["bar", "spam", "baz"]),
                     ],
@@ -511,7 +699,7 @@ foo ha"
                     .trim(),
                     file_extension: "md",
                     prepare_events: &[],
-                    events: keys!("w l q ."),
+                    events: keys!("t l q ."),
                     expectations: &[
                         CurrentSelectedTexts(&["beta", "gamma", "omega", "zeta"]),
                     ],
@@ -525,7 +713,7 @@ foo ha"
                     .trim(),
                     file_extension: "md",
                     prepare_events: keys!("/ z enter"),
-                    events: keys!("w h q ,"),
+                    events: keys!("t h q ,"),
                     expectations: &[
                         CurrentSelectedTexts(&["alpha","beta", "gamma", "omega"]),
                     ],
@@ -1007,7 +1195,7 @@ pub(crate) fn from_text(language: Option<tree_sitter::Language>, text: &str) -> 
                     file_extension: "rs",
                     prepare_events: &[],
                     events: keys!(
-                        "/ y x enter s q q b n v s ( i S o m e esc s b n n T q o"
+                        "/ y x enter s q q j n v s ( i S o m e esc s j n n ^ q o"
                     ),
                     expectations: &[],
                     terminal_height: None,
@@ -1049,7 +1237,7 @@ fn syntax_node() -> RecipeGroup {
                 description: "Select a syntax node (JSON)",
                 content: "[{\"x\": 123}, true, {\"y\": {}}]".trim(),
                 file_extension: "json",
-                prepare_events: keys!("^ l"),
+                prepare_events: keys!("z l"),
                 events: keys!("s"),
                 expectations: &[CurrentSelectedTexts(&["{\"x\": 123}"])],
                 terminal_height: None,
@@ -1060,8 +1248,8 @@ fn syntax_node() -> RecipeGroup {
                 description: "Navigate sibling nodes via Next/Previous/First/Last movement",
                 content: "[{\"x\": 123}, true, {\"y\": {}}]".trim(),
                 file_extension: "json",
-                prepare_events: keys!("^ l"),
-                events: keys!("s n n N N . ,"),
+                prepare_events: keys!("z l"),
+                events: keys!("s n n b b . ,"),
                 expectations: &[CurrentSelectedTexts(&["{\"x\": 123}"])],
                 terminal_height: None,
                 similar_vim_combos: &[],
@@ -1079,22 +1267,22 @@ fn syntax_node() -> RecipeGroup {
                 only: false,
             },
             Recipe {
-                description: "Expand selection / Select Parent",
+                description: "Select Parent",
                 content: "[{\"x\": 123}, true, {\"y\": {}}]".trim(),
                 file_extension: "json",
-                prepare_events: keys!("^ l l"),
-                events: keys!("S t t t t"),
+                prepare_events: keys!("z l l"),
+                events: keys!("S k k k k"),
                 expectations: &[CurrentSelectedTexts(&["[{\"x\": 123}, true, {\"y\": {}}]"])],
                 terminal_height: None,
                 similar_vim_combos: &[],
                 only: false,
             },
             Recipe {
-                description: "Shrink selection / Select First-Child",
+                description: "Select First-Child",
                 content: "[{\"x\": 123}, true, {\"y\": {}}]".trim(),
                 file_extension: "json",
                 prepare_events: keys!("s"),
-                events: keys!("s b b b b"),
+                events: keys!("s j j j j"),
                 expectations: &[CurrentSelectedTexts(&["x"])],
                 terminal_height: None,
                 similar_vim_combos: &[],
@@ -1209,22 +1397,6 @@ And by opposing end them. To die—to sleep,",
             only: false,
         },
         Recipe {
-            description: "Delete lines",
-            content: "
-To be, or not to be?
-That, is the question.
-Why?
-"
-            .trim(),
-            file_extension: "md",
-            prepare_events: &[],
-            events: keys!("e d d"),
-            expectations: &[CurrentSelectedTexts(&["Why?"])],
-            terminal_height: None,
-            similar_vim_combos: &["d d", "d j"],
-            only: false,
-        },
-        Recipe {
             description: "Insert at the beginning of line",
             content: "  hat is that?",
             file_extension: "md",
@@ -1247,18 +1419,18 @@ Why?
             only: false,
         },
         Recipe {
-            description: "Word movement",
+            description: "Token movement",
             content: "hello-world camelCase snake_case",
             file_extension: "md",
             prepare_events: &[],
-            events: keys!("w l l"),
+            events: keys!("t l l"),
             expectations: &[CurrentSelectedTexts(&["snake_case"])],
             terminal_height: None,
             similar_vim_combos: &["w", "e", "b"],
             only: false,
         },
         Recipe {
-            description: "Sub words movement",
+            description: "Word movement",
             content: "
 camelCase
 hello_world
@@ -1266,30 +1438,8 @@ hello_world
             .trim(),
             file_extension: "md",
             prepare_events: &[],
-            events: keys!("W j l k h"),
+            events: keys!("w j l k h"),
             expectations: &[CurrentSelectedTexts(&["camel"])],
-            terminal_height: None,
-            similar_vim_combos: &[],
-            only: false,
-        },
-        Recipe {
-            description: "Delete subwords (forward)",
-            content: "camelCase snake_case".trim(),
-            file_extension: "md",
-            prepare_events: &[],
-            events: keys!("W d d"),
-            expectations: &[CurrentSelectedTexts(&["snake"])],
-            terminal_height: None,
-            similar_vim_combos: &[],
-            only: false,
-        },
-        Recipe {
-            description: "Delete subwords (backward)",
-            content: "camelCase snake_case".trim(),
-            file_extension: "md",
-            prepare_events: &[],
-            events: keys!("W . D D"),
-            expectations: &[CurrentSelectedTexts(&["snake"])],
             terminal_height: None,
             similar_vim_combos: &[],
             only: false,
@@ -1299,98 +1449,10 @@ hello_world
             content: "camelCase".trim(),
             file_extension: "md",
             prepare_events: &[],
-            events: keys!("W d u U"),
+            events: keys!("w d u U"),
             expectations: &[CurrentComponentContent("Case")],
             terminal_height: None,
             similar_vim_combos: &["u", "ctrl+r"],
-            only: false,
-        },
-        Recipe {
-            description: "Surround",
-            content: "hello world".trim(),
-            file_extension: "md",
-            prepare_events: &[],
-            events: keys!("e v s ("),
-            expectations: &[CurrentComponentContent("(hello world)")],
-            terminal_height: None,
-            similar_vim_combos: &[],
-            only: false,
-        },
-        Recipe {
-            description: "Delete Surround",
-            content: "(hello world)".trim(),
-            file_extension: "md",
-            prepare_events: keys!("^ l"),
-            events: keys!("v d ("),
-            expectations: &[CurrentComponentContent("hello world")],
-            terminal_height: None,
-            similar_vim_combos: &[],
-            only: false,
-        },
-        Recipe {
-            description: "Change Surround",
-            content: "(hello world)".trim(),
-            file_extension: "md",
-            prepare_events: keys!("^ l"),
-            events: keys!("v c ( {"),
-            expectations: &[CurrentComponentContent("{hello world}")],
-            terminal_height: None,
-            similar_vim_combos: &[],
-            only: false,
-        },
-        Recipe {
-            description: "Select Inside Enclosures",
-            content: "(hello world)".trim(),
-            file_extension: "md",
-            prepare_events: keys!("^ l"),
-            events: keys!("v i ("),
-            expectations: &[CurrentSelectedTexts(&["hello world"])],
-            terminal_height: None,
-            similar_vim_combos: &["v i ("],
-            only: false,
-        },
-        Recipe {
-            description: "Select Around Enclosures",
-            content: "(hello world)".trim(),
-            file_extension: "md",
-            prepare_events: keys!("^ l"),
-            events: keys!("v a ("),
-            expectations: &[CurrentSelectedTexts(&["(hello world)"])],
-            terminal_height: None,
-            similar_vim_combos: &["v a ("],
-            only: false,
-        },
-        Recipe {
-            description: "Extend selection (Word)",
-            content: "foo bar spam pi".trim(),
-            file_extension: "md",
-            prepare_events: &[],
-            events: keys!("w v l l h"),
-            expectations: &[CurrentSelectedTexts(&["foo bar"])],
-            terminal_height: None,
-            similar_vim_combos: &[],
-            only: false,
-        },
-        Recipe {
-            description: "Extend selection (Syntax Node)",
-            content: "[{\"x\": 123}, true, {\"y\": {}}]".trim(),
-            file_extension: "json",
-            prepare_events: keys!("^ l"),
-            events: keys!("s v n n N"),
-            expectations: &[CurrentSelectedTexts(&["{\"x\": 123}, true"])],
-            terminal_height: None,
-            similar_vim_combos: &[],
-            only: false,
-        },
-        Recipe {
-            description: "Extend selection (Switch Direction)",
-            content: "foo bar spam baz tim".trim(),
-            file_extension: "md",
-            prepare_events: &[],
-            events: keys!("w l l v l o h o l"),
-            expectations: &[CurrentSelectedTexts(&["bar spam baz tim"])],
-            terminal_height: None,
-            similar_vim_combos: &[],
             only: false,
         },
         Recipe {
@@ -1430,7 +1492,7 @@ foo bar spam",
             description: "Move the first two elements to the last",
             content: "[{\"a\": b}, \"c\", [], {}]".trim(),
             file_extension: "json",
-            prepare_events: keys!("^ l"),
+            prepare_events: keys!("z l"),
             events: keys!("s v n y d . p"),
             expectations: &[CurrentComponentContent("[[], {}, {\"a\": b}, \"c\"]")],
             terminal_height: None,
@@ -1453,7 +1515,7 @@ foo bar spam",
             content: "if(condition) { x(bar(baz)) } else { 'hello world' }".trim(),
             file_extension: "js",
             prepare_events: keys!("/ x enter"),
-            events: keys!("s T r"),
+            events: keys!("s ^ r"),
             expectations: &[CurrentComponentContent("x(bar(baz))")],
             terminal_height: Some(7),
             similar_vim_combos: &[],
@@ -1507,29 +1569,11 @@ And drop on the deck and flop like a fish?
             only: false,
         },
         Recipe {
-            description: "Replace cut",
-            content: "
-foo(bar, 1 + 1, spam)
-3 * 10
-"
-            .trim(),
-            file_extension: "js",
-            prepare_events: &[],
-            events: keys!("s y n R b n b n r"),
-            expectations: &[CurrentComponentContent(
-                "foo(bar, 1 + 1, spam)
-foo(bar, 3 * 10, spam)",
-            )],
-            terminal_height: None,
-            similar_vim_combos: &["p"],
-            only: false,
-        },
-        Recipe {
             description: "Invert nesting (JSX)",
             content: "<Parent><Child><Grandson/></Child></Parent>".trim(),
             file_extension: "js",
-            prepare_events: keys!("s b n b n"),
-            events: keys!("y t R t R b n r"),
+            prepare_events: keys!("s j n j n"),
+            events: keys!("y k R k R j n r"),
             expectations: &[CurrentComponentContent(
                 "<Child><Parent><Grandson/></Parent></Child>",
             )],
@@ -1542,7 +1586,7 @@ foo(bar, 3 * 10, spam)",
             content: "foo(bar(yo, spam(baz), baz), bomb)".trim(),
             file_extension: "js",
             prepare_events: keys!("/ s enter s"),
-            events: keys!("y t t R t t R b n b n r"),
+            events: keys!("y k k R k k R j n j n r"),
             expectations: &[CurrentComponentContent(
                 "bar(yo, foo(spam(baz), bomb), baz)",
             )],
@@ -1562,60 +1606,12 @@ foo(bar, 3 * 10, spam)",
             only: false,
         },
         Recipe {
-            description: "Collapse selection (Start)",
-            content: "foo bar spam".trim(),
-            file_extension: "js",
-            prepare_events: &[],
-            events: keys!("e ^"),
-            expectations: &[CurrentSelectedTexts(&["f"])],
-            terminal_height: None,
-            similar_vim_combos: &[],
-            only: false,
-        },
-        Recipe {
             description: "Select from current selection until end of line",
             content: "foo bar spam".trim(),
             file_extension: "js",
-            prepare_events: keys!("w l"),
+            prepare_events: keys!("t l"),
             events: keys!("v e $"),
             expectations: &[CurrentSelectedTexts(&["bar spam"])],
-            terminal_height: None,
-            similar_vim_combos: &[],
-            only: false,
-        },
-        Recipe {
-            description: "Select from current selection until beginning of line",
-            content: "foo bar spam".trim(),
-            file_extension: "js",
-            prepare_events: keys!("w l"),
-            events: keys!("v e ^"),
-            expectations: &[CurrentSelectedTexts(&["foo bar"])],
-            terminal_height: None,
-            similar_vim_combos: &[],
-            only: false,
-        },
-        Recipe {
-            description: "Paste with automatic gap insertion",
-            content: "
-foo bar
-spam baz
-"
-            .trim(),
-            file_extension: "md",
-            prepare_events: &[],
-            events: keys!("e y p"),
-            expectations: &[CurrentComponentContent("foo bar\nfoo bar\nspam baz")],
-            terminal_height: None,
-            similar_vim_combos: &[],
-            only: false,
-        },
-        Recipe {
-            description: "Paste without automatic gap insertion",
-            content: "foo bar".trim(),
-            file_extension: "md",
-            prepare_events: &[],
-            events: keys!("e y $ p"),
-            expectations: &[CurrentComponentContent("foo barfoo bar")],
             terminal_height: None,
             similar_vim_combos: &[],
             only: false,
