+++ src/selection_mode/mod.rs
@@ -5,7 +5,7 @@ pub(crate) mod diagnostic;
 pub(crate) mod git_hunk;
 pub(crate) mod mark;
 pub(crate) mod naming_convention_agnostic;
-pub(crate) mod token;
+pub(crate) mod syntax_token;
 
 pub(crate) mod top_node;
 
@@ -13,8 +13,8 @@ pub(crate) mod line_full;
 pub(crate) mod line_trimmed;
 pub(crate) mod local_quickfix;
 pub(crate) mod regex;
-pub(crate) mod subword;
 pub(crate) mod syntax_node;
+pub(crate) mod token;
 pub(crate) mod word;
 pub(crate) use self::regex::Regex;
 pub(crate) use ast_grep::AstGrep;
@@ -30,8 +30,8 @@ pub(crate) use mark::Mark;
 pub(crate) use naming_convention_agnostic::NamingConventionAgnostic;
 use position_pair::ParsedChar;
 use std::ops::Range;
-pub(crate) use subword::Subword;
 pub(crate) use syntax_node::SyntaxNode;
+pub(crate) use syntax_token::SyntaxToken;
 pub(crate) use token::Token;
 pub(crate) use top_node::TopNode;
 pub(crate) use word::Word;
@@ -198,7 +198,6 @@ pub trait SelectionMode {
             Movement::Down => convert(self.down(params)),
             Movement::ToParentLine => convert(self.to_parent_line(params)),
             Movement::Expand => self.expand(params),
-            Movement::Shrink => self.shrink(params),
             Movement::Next => convert(self.next(params)),
             Movement::Previous => convert(self.previous(params)),
             Movement::DeleteBackward => convert(self.delete_backward(params)),
@@ -360,10 +359,6 @@ pub trait SelectionMode {
         Ok(Some(ApplyMovementResult::from_selection(
             selection.clone().set_range(range),
         )))
-    }
-
-    fn shrink(&self, _: SelectionModeParams) -> anyhow::Result<Option<ApplyMovementResult>> {
-        Ok(None)
     }
 
     fn next(&self, params: SelectionModeParams) -> anyhow::Result<Option<Selection>> {
