+++ src/selection_mode/syntax_node.rs
@@ -2,7 +2,7 @@ use itertools::Itertools;
 
 use crate::selection_mode::ApplyMovementResult;
 
-use super::{ByteRange, SelectionMode, Token, TopNode};
+use super::{ByteRange, SelectionMode, SyntaxToken, TopNode};
 
 pub(crate) struct SyntaxNode {
     /// If this is true:
@@ -19,7 +19,7 @@ impl SelectionMode for SyntaxNode {
         if self.coarse {
             TopNode.iter(params)
         } else {
-            Token.iter(params)
+            SyntaxToken.iter(params)
         }
     }
     fn expand(
@@ -35,12 +35,7 @@ impl SelectionMode for SyntaxNode {
         self.select_vertical(params, false)
             .map(|result| result.map(|result| result.selection))
     }
-    fn shrink(
-        &self,
-        params: super::SelectionModeParams,
-    ) -> anyhow::Result<Option<ApplyMovementResult>> {
-        self.select_vertical(params, false)
-    }
+
     fn up(
         &self,
         params: super::SelectionModeParams,
@@ -298,13 +293,13 @@ mod test_syntax_node {
 
             let parent_text = buffer.slice(&parent_range).unwrap();
             assert_eq!(parent_text, "{z}");
-            let selection = SyntaxNode { coarse }.shrink(SelectionModeParams {
+            let selection = SyntaxNode { coarse }.down(SelectionModeParams {
                 buffer: &buffer,
                 current_selection: &Selection::new(parent_range),
                 cursor_direction: &crate::components::editor::Direction::Start,
             });
 
-            let child_range = selection.unwrap().unwrap().selection.range();
+            let child_range = selection.unwrap().unwrap().range();
 
             let child_text = buffer.slice(&child_range).unwrap();
             assert_eq!(child_text, expected_child);
